Fix bug of AlertDialog displaying wrong player as winner
Order of players in the score screen is swapped after each frame
to change the player who takes the first shot. Currently after
the frame the name of the winner provided to AlertDialog doesn't
take this into account, and thus the wrong player is displayed
as winner after every second frame.

Amend the logic responsible providing the name of the winner to
AlertDialog to take possible reversed display order into
consideration.