Add indicator to know that free ball was potted
Let's imagine that player A does a foul and player B gets free
ball. B is on reds, and chooses blue as the free ball. B pots
the blue, getting 1 point. Currently the break screen shows
this as a normal red ball, and it can be confusing.

Add "firstWasFreeBall"-variable to BreakUi, and add logic to
show an indicator (letter 'F') inside a potted free ball. This
is handled identically to redsRemaining by Ball-composable.