Add check for update from Play Store
Currently if users have disabled automatic updates, they might
have no idea if an update has been released to this app.

Add an automatic check via Play api to MainActivity, which
prompts to update the app if an important (at least level 1)
update is available. Importance is set individually to each
update with Play Console api, and it ranges from 0 to 5.